Implement std::IN as a derivative of std::=
The membership operators, `std::IN` and `std::NOT IN` must be applicable
to the same domain as `std::=`, since `[NOT] IN` is basically an
element-wise application of `=`.  Blindly copying all the `=`
definitions would be tedious now and in the future, so introduce a
mechanism to declare an operator as a "derivative" of another operator,
which would tell the compiler to use the origin operator when doing
call resolution.  For simplicity, derivative operators cannot be
polymorphic.

Also fixes an issue with tuple inputs to `IN`, where the right operand was
erroneously passed as a TupleVar instead of a value.