Add sys::current_permissions. (#9054)
This provides a simple way for a user to check all of their permissions.

Previously, the only way for a non-superuser to do this was to select
all permission names, then check each one. This would have to be done
per branch, since they may have different permissions, but would also
miss any permissions which were not yet created.