Support repair patches conditional on extension presence (#9078)
We keep having situations where we need to run a repair patch just to
fix the AI extension. Because of the despiséd annotation copying, we
need to do this every time we fix a model parameter or add a new one,
and we've also needed it because of other screwups (like #9041).