Check for explicit id write presence in a slightly more resilient way (#9090)
Currently we distinguish an explicit id write from an implicitly
injected id or a default value by whether a span is present.

This is kind of fragile, especially in the implicit case,
so also check whether an expression is present.

This will fix #9089 if back-ported to 6.x. That bug was already
accidentally fixed by #8531.