Minimize FlatSchema interface (#9016)
Goal of this PR is twofold:
- improve performance of `s_schema.Schema` data structure,
- minimize its interface,

The idea is to move as much behavior out of the data structure itself
into helper functions around it. For example:
- `def get()` performs name resolution (which includes application of
module aliases, implicit prefixing with `std` and a few more things). If
this would be replaced with a simpler:
  - `def get_by_qualified_name(sn.Name)` and
  - `get_by_shortname(str, SchemaClass)` and
  - `get_by_globalname(str, SchemaClass)`
Now most calls of `get` could be replaced with more specific, faster
function.