Fix an interaction between aliases and optional globals with a default (#9109)
Optional globals with a default get turned into *two* parameters, one
for the value and one for whether they were set. This is to
distinguish between the cases of setting to `{}` and not setting
anything.

The "unused parameter" handling wasn't accounting for this so we would
pass more arguments than the query expected in cases where a global
got seen during compilation but didn't end up in the final query.