Revert "Fix half-dunder pointer names (#9100)" (#9115)
This reverts commit 9c9594f85beffa677c936d3da060d742c46bbad7.

Fixing half-dunder names has the problem that it leaves any database
still using them (via an inplace-upgrade or minor version upgrade) in
a pretty broken state, where some of the postgres columns may have
half-dunder names, but now there is now way to recover the situation:
it won't be possible to rename or delete the pointers.

In order to fix this, we'll probably need migration code that can run
during an inplace upgrade.